Update button click handling and view constants usage
Enhanced button click handling in 'SideBarViewController' and introduced a new usage of view constants from 'Views' class. Now, the button click resets the content area view, updates the style of the clicked button and sets the content and anchor based on the clicked button. Moreover, instead of hardcoding the path, 'EDIT_VIEW' constant from the 'Views' class is being used.

--Improved performance dramatically