Add unzip utility method in FileUtils
A utility method for unzipping files has been added to the FileUtils class. The method takes an InputStream of the zip file and a destination directory as parameters. It extracts the contents of the given zip file to the specified directory, throwing an IOException in case of an I/O error.